Rework init_checkbox and use gsettings bind method
We now bind the CheckButton active property directly to the key in gsettings.
This makes the key and the widget update when the other changes.

Connect the CheckButton's toggle signal to a new function that handles the
the restarting of maximus. This code previously resided in the set_bool
function where it does not belong.